utils symlinking + install pip packages via script + minor fixes
