+++ scripts/init.sh
@@ -1,3 +1,6 @@
 # Script to Initialize the project
 pyenv local 3.9.12
-pip install -r requirements.txt
+bash scripts/createAllUtilsSymlink.sh
+checks_dir=src/checks/
+pip install --upgrade pip
+find $checks_dir -name requirements.txt -exec pip install -r {} \;
