+++ src/checks/conciseness/conciseness.py
