+++ src/checks/conciseness/prompt.py
