+++ src/checks/conciseness/tests/conciseness_test.py
