+++ src/checks/conciseness/tests/promptfoo/promptfooconfig.yaml
