+++ src/checks/conciseness/tests/promptfoo/prompts.txt
