+++ src/checks/conciseness/utils
@@ -1,0 +1,1 @@
+/workspace/check-functions/src/utils
