+++ src/checks/non_committal/prompt.py
